SECURITY + CAPTCHA

+++ src/main/java/org/diplom/blog/dto/UserDto.java
@@ -7,11 +7,8 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.Accessors;
 
-//@Getter
-//@Setter
 @Data
 @Accessors(chain = true)
-//@NoArgsConstructor
 @JsonInclude(value = JsonInclude.Include.NON_NULL)
 public class UserDto {
     private Long id;
@@ -20,6 +17,6 @@ public class UserDto {
     private Integer removePhoto;
     private String email;
     private Boolean moderation;
-    private Integer moderationCount;
+    private Long moderationCount;
     private Boolean settings;
 }
